EVG-17874: Add ability to set "activate" value for tasks.
This value was introduced a couple of years ago, and we need to
update the v1 version of shrub.py so that Genny can generate
tasks that don't auto activate.

Note that undefined/null is different from both true and false for
activate, since in a task if activate is undefined it will try to
get the setting from the version or batchtime. Because of this,
there are some code changes to make sure this setting works
properly if set to any of null, false, or true.