EVG-17874: Updates to get project to work with modern Python on M1.
Need to get rid of the typed-ast module so we don't have to compile
from source on M1. The pytest plugins are all defunct and don't
have new enough versions to avoid typed ast, so this change
updates the minimum python version to 3.7 to use new black that
doesn't depend on it, and manually runs black and flake8 instead
of using the pytest plugins.